Simplify the same code
